Several fixes for compiling with CLang
git-svn-id: file:///home/michael/svn/yuni/trunk@1383 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/io/directory/info/platform.cpp
@@ -136,7 +136,7 @@ namespace Directory
 					if (0 != (flags & Yuni::Core::IO::Directory::Info::itFile))
 					{
 						isFolder = false;
-						size = s.st_size;
+						size = static_cast<uint64>(s.st_size);
 						return true;
 					}
 				}
